+++ src/main/java/nextstep/session/domain/constraint/SessionConstraint.java
@@ -1,4 +1,4 @@
-package nextstep.session.domain.payment;
+package nextstep.session.domain.constraint;
 
 public class SessionConstraint {
