+++ src/test/java/nextstep/payments/domain/PaymentsTest.java
@@ -4,7 +4,7 @@ import nextstep.courses.domain.Course;
 import nextstep.session.domain.Session;
 import nextstep.session.domain.SessionPeriod;
 import nextstep.session.domain.image.SessionImage;
-import nextstep.session.domain.payment.SessionConstraint;
+import nextstep.session.domain.constraint.SessionConstraint;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
